[Enhance #169] View Global Activity Log
