+++ app/components/badge/index.ts
@@ -1,0 +1,2 @@
+export { default as BadgeCondition } from './badge-condition';
+export { default as BadgeState } from './badge-state';
