+++ app/components/button/index.ts
@@ -1,0 +1,3 @@
+export { default as ButtonCopy } from './button-copy';
+export { default as ButtonDeleteAction } from './button-delete-action';
+export { default as ButtonLoading } from './button-loading';
