+++ app/components/date/index.ts
@@ -1,0 +1,2 @@
+export { default as DateFormatter } from './date-formatter';
+export { DateRangePicker } from './date-range-picker';
