+++ app/components/dialog/index.ts
@@ -1,0 +1,1 @@
+export { default as DialogConfirm } from './dialog-confirm';
