+++ app/components/list/index.ts
@@ -1,0 +1,1 @@
+export { default as ListActivity } from './list-activity';
