+++ app/components/tooltip/index.ts
@@ -1,0 +1,1 @@
+export { default as Tooltip } from './tooltip';
