+++ app/modules/loki/loki-client.ts
@@ -103,6 +103,7 @@ export async function executeLokiQuery(
       timerange: response.timerange,
     });
 
+    console.log('response', JSON.stringify(response.logs));
     return response;
   } catch (error) {
     logger.error('Loki query failed', {
