+++ app/resources/request/client/project.request.ts
@@ -1,5 +1,4 @@
 import { apiRequestClient } from '@/modules/axios/axios.client';
-import { ProjectActivityListResponseSchema } from '@/resources/schemas/activity.schema';
 import { ListQueryParams } from '@/resources/schemas/common.schema';
 import { HTTPProxyListResponseSchema } from '@/resources/schemas/httpproxy.schema';
 import { ProjectListResponseSchema } from '@/resources/schemas/project.schema';
@@ -35,23 +34,4 @@ export const projectDeleteMutation = (projectName: string) => {
     method: 'DELETE',
     url: `/apis/resourcemanager.miloapis.com/v1alpha1/projects/${projectName}`,
   }).execute();
-};
-
-export const projectActivityListQuery = (projectName: string, params?: ListQueryParams) => {
-  return apiRequestClient({
-    method: 'GET',
-    url: '',
-    baseURL: '/api/activity',
-    params: {
-      // Only Write operations are logged
-      actions: 'create,update,patch,delete,deletecollection',
-      project: projectName,
-      ...(params?.limit && { limit: params.limit }),
-      ...(params?.filters?.start && { start: params.filters.start }), // Already in nanoseconds
-      ...(params?.filters?.end && { end: params.filters.end }), // Already in nanoseconds
-      ...(params?.search && { q: params.search }),
-    },
-  })
-    .output(ProjectActivityListResponseSchema)
-    .execute();
 };
