+++ app/routes/organization/detail/index.tsx
@@ -1,7 +1,7 @@
 import type { Route } from './+types/index';
-import BadgeState from '@/components/badge-state';
-import CopyButton from '@/components/copy-button';
-import DateFormatter from '@/components/date-formatter';
+import { BadgeState } from '@/components/badge';
+import { ButtonCopy } from '@/components/button';
+import { DateFormatter } from '@/components/date';
 import { Text, Title } from '@/components/typography';
 import { Card, CardContent } from '@/modules/shadcn/ui/card';
 import { Table, TableBody, TableCell, TableRow } from '@/modules/shadcn/ui/table';
