+++ app/routes/user/detail/layout.tsx
@@ -1,5 +1,5 @@
 import type { Route } from './+types/layout';
-import DeleteActionButton from '@/components/delete-action-button';
+import { ButtonDeleteAction } from '@/components/button';
 import { SubLayout } from '@/components/sub-layout';
 import { authenticator } from '@/modules/auth';
 import { toast } from '@/modules/toast';
@@ -42,7 +42,7 @@ export default function Layout() {
   return (
     <SubLayout>
       <SubLayout.ActionBar>
-        <DeleteActionButton
+        <ButtonDeleteAction
           itemType="User"
           description={t`Are you sure you want to delete user "${data.spec.givenName} ${data.spec.familyName}"? This action cannot be undone.`}
           onConfirm={handleDeleteUser}
